added new version of dimensional doors
1. Fixed a bug in converting old save files (from before MC 1.6) to the
current save format
2. Fixed a crash that would occur when players tried to place doors
from certain mods
3. Stopped Monoliths from teleporting players while in Limbo
4. Increased the maximum stack size of Golden Doors to 16 items
5. Changed the crafting recipes for most DD items to use Ender Pearls,
as per discussions on this thread. The items that still use Stable
Fabric are Dimensional Doors, Golden Dimensional Doors, Rift Blades,
and Stabilized Rift Signatures.
6. Changed the crafting recipe for Stabilized Rift Signatures to 4 Iron
Ingots and a Stable Fabric
7. Changed the Stabilized Rift Signature so that it consumes Ender
Pearls instead of Stable Fabric
8. Changed the Stabilized Rift Signature so that "redirection"
operations can be done for free. Suppose that you have two rifts A and
B which are connected together, with A as the main rift for an SRS. If
you create another a rift C with the SRS, then rift A will point to
rift C. You can now click on the same place where rift B was created to
redirect rift A to link to rift B without consuming Ender Pearls. This
is fair because no new rifts are being created in the process.
9. Fixed a bug where the Enderman spawning setting for rifts was being
ignored
10. Added support for using a Stabilized Rift Signature from a
Dispenser. The SRS will only activate to perform a free redirection as
described in change #8.
11. Players will now recover their food levels when respawned in Limbo
12. Added the Universal Limbo setting to the World config options. When
enabled, players who die in any dimension (except Limbo) will be sent
to Limbo. This is disabled by default. [Requested by Mr_Turing]
13. Fixed an issue where Golden Dimensional Doors would request
duplicate chunkloader tickets each time the server rebooted. This would
not actually load chunks repeatedly, but Opis would detect the chunks
as loaded repeatedly for each ticket, and it would report ridiculous
numbers of loaded chunks in pockets with golden doors.
14. Fixed the calculation for pocket boundaries so that Golden
Dimensional Doors load the correct chunks for a pocket
15. Stopped Golden Dimensional Doors from being able to chunkload a
pocket if they're outside the horizontal bounds of the pocket